// src/types/StopInfo.ts
export interface StopInformation {
    id: number;
    stop_id: number;
    stop_code: string;
    stop_name: string;
    stop_lat: number;
    stop_lon: number;
    zone_id: number;
    alias: string;
    stop_area: string;
    stop_desc: string;
    lest_x: number;
    lest_y: number;
    zone_name: string;
    authority: string;
}